Add slash command support for existing commands

+++ src/main/java/net/jellyo/jmusicbot/commands/MusicCommand.java
@@ -50,9 +50,10 @@ public abstract class MusicCommand extends Command
         TextChannel tchannel = settings.getTextChannel(event.getGuild());
         if(tchannel!=null && !event.getTextChannel().equals(tchannel))
         {
-            try 
+            try
             {
-                event.getMessage().delete().queue();
+                if(event.getMessage()!=null)
+                    event.getMessage().delete().queue();
             } catch(PermissionException ignore){}
             event.replyInDm(event.getClient().getError()+" You can only use that command in "+tchannel.getAsMention()+"!");
             return;
